fingerprinti creator mad more general as a recording tool
git-svn-id: svn://svn.code.sf.net/p/jsidplay2/code/trunk@4889 05cac4c1-a8e4-46c3-9d4b-80028926d4ef

+++ jsidplay2/src/main/java/ui/tools/RecordingTool.java
@@ -1,7 +1,6 @@
 package ui.tools;
 
-import static libsidplay.config.IWhatsSidSystemProperties.FINGERPRINTING_AWAIT_TERMINATION;
-import static libsidplay.config.IWhatsSidSystemProperties.FINGERPRINTING_MAX_THREADS;
+import static libsidplay.config.IWhatsSidSystemProperties.AWAIT_TERMINATION;
 
 import java.io.File;
 import java.io.IOException;
@@ -33,10 +32,12 @@ import libsidutils.fingerprinting.FingerPrinting;
 import libsidutils.fingerprinting.ini.IniFingerprintConfig;
 import libsidutils.siddatabase.SidDatabase;
 import sidplay.Player;
+import sidplay.audio.AudioDriver;
+import sidplay.filefilter.AudioTuneFileFilter;
+import sidplay.filefilter.VideoTuneFileFilter;
 import sidplay.ini.IniConfig;
 import sidplay.ini.converter.FileToStringConverter;
 import sidplay.player.DebugUtil;
-import ui.common.filefilter.TuneFileFilter;
 import ui.entities.PersistenceProperties;
 import ui.entities.whatssid.service.WhatsSidService;
 import ui.tools.audio.WhatsSidDriver;
@@ -53,36 +54,32 @@ import ui.tools.audio.WhatsSidDriver;
  * @author ken
  *
  */
-@Parameters(resourceBundle = "ui.tools.FingerPrintingCreator")
-public class FingerPrintingCreator {
+@Parameters(resourceBundle = "ui.tools.RecordingTool")
+public class RecordingTool {
 
 	static {
 		DebugUtil.init();
 	}
 
-	private static final TuneFileFilter TUNE_FILE_FILTER = new TuneFileFilter();
+	private static final AudioTuneFileFilter AUDIO_TUNE_FILE_FILTER = new AudioTuneFileFilter();
+	private static final VideoTuneFileFilter VIDEO_TUNE_FILE_FILTER = new VideoTuneFileFilter();
 
 	@Parameter(names = { "--help", "-h" }, descriptionKey = "USAGE", help = true, order = 10000)
 	private Boolean help = Boolean.FALSE;
 
-	@Parameter(names = {
-			"--whatsSIDDatabaseDriver" }, descriptionKey = "WHATSSID_DATABASE_DRIVER", required = true, order = 10001)
+	@Parameter(names = { "--whatsSIDDatabaseDriver" }, descriptionKey = "WHATSSID_DATABASE_DRIVER", order = 10001)
 	private String whatsSidDatabaseDriver;
 
-	@Parameter(names = {
-			"--whatsSIDDatabaseUrl" }, descriptionKey = "WHATSSID_DATABASE_URL", required = true, order = 10002)
+	@Parameter(names = { "--whatsSIDDatabaseUrl" }, descriptionKey = "WHATSSID_DATABASE_URL", order = 10002)
 	private String whatsSidDatabaseUrl;
 
-	@Parameter(names = {
-			"--whatsSIDDatabaseUsername" }, descriptionKey = "WHATSSID_DATABASE_USERNAME", required = true, order = 10003)
+	@Parameter(names = { "--whatsSIDDatabaseUsername" }, descriptionKey = "WHATSSID_DATABASE_USERNAME", order = 10003)
 	private String whatsSidDatabaseUsername;
 
-	@Parameter(names = {
-			"--whatsSIDDatabasePassword" }, descriptionKey = "WHATSSID_DATABASE_PASSWORD", required = true, order = 10004)
+	@Parameter(names = { "--whatsSIDDatabasePassword" }, descriptionKey = "WHATSSID_DATABASE_PASSWORD", order = 10004)
 	private String whatsSidDatabasePassword;
 
-	@Parameter(names = {
-			"--whatsSIDDatabaseDialect" }, descriptionKey = "WHATSSID_DATABASE_DIALECT", required = true, order = 10005)
+	@Parameter(names = { "--whatsSIDDatabaseDialect" }, descriptionKey = "WHATSSID_DATABASE_DIALECT", order = 10005)
 	private String whatsSidDatabaseDialect;
 
 	@Parameter(names = { "--createIni" }, descriptionKey = "CREATE_INI", arity = 1, order = 10006)
@@ -91,8 +88,14 @@ public class FingerPrintingCreator {
 	@Parameter(names = { "--deleteAll" }, descriptionKey = "DELETE_ALL", arity = 1, order = 10007)
 	private Boolean deleteAll = Boolean.FALSE;
 
+	@Parameter(names = { "--fingerprinting" }, descriptionKey = "FINGERPRINTING", arity = 1, order = 10008)
+	private Boolean fingerprinting = Boolean.FALSE;
+
+	@Parameter(names = { "--maxThreads" }, descriptionKey = "MAX_THREADS", order = 10009)
+	private Integer maxThreads = 8;
+
 	@Parameter(names = {
-			"--previousDirectory" }, descriptionKey = "PREVIOUS_DIRECTORY", converter = FileToStringConverter.class, order = 10008)
+			"--previousDirectory" }, descriptionKey = "PREVIOUS_DIRECTORY", converter = FileToStringConverter.class, order = 10010)
 	private File previousDirectory;
 
 	@Parameter(description = "directory", converter = FileToStringConverter.class)
@@ -110,7 +113,7 @@ public class FingerPrintingCreator {
 	private volatile boolean quit;
 
 	private void execute(String[] args) {
-		ExecutorService executor = Executors.newFixedThreadPool(FINGERPRINTING_MAX_THREADS);
+		ExecutorService executor = Executors.newFixedThreadPool(maxThreads);
 		try {
 			JCommander commander = JCommander.newBuilder().addObject(this).programName(getClass().getName()).build();
 			commander.parse(args);
@@ -124,7 +127,9 @@ public class FingerPrintingCreator {
 				System.out.println("Parameter --hvsc must be present!");
 				System.exit(1);
 			}
-			config.getAudioSection().setSamplingRate(SamplingRate.VERY_LOW);
+			if (fingerprinting) {
+				config.getAudioSection().setSamplingRate(SamplingRate.VERY_LOW);
+			}
 			config.getSidplay2Section().setDefaultPlayLength(180);
 			config.getSidplay2Section().setEnableDatabase(true);
 			config.getSidplay2Section().setSingle(false);
@@ -134,19 +139,20 @@ public class FingerPrintingCreator {
 				previousSidDatabase = new SidDatabase(previousDirectory);
 			}
 
-			entityManagerFactory = Persistence.createEntityManagerFactory(PersistenceProperties.WHATSSID_DS,
-					new PersistenceProperties(whatsSidDatabaseDriver, whatsSidDatabaseUrl, whatsSidDatabaseUsername,
-							whatsSidDatabasePassword, whatsSidDatabaseDialect));
+			if (whatsSidDatabaseDriver != null) {
+				entityManagerFactory = Persistence.createEntityManagerFactory(PersistenceProperties.WHATSSID_DS,
+						new PersistenceProperties(whatsSidDatabaseDriver, whatsSidDatabaseUrl, whatsSidDatabaseUsername,
+								whatsSidDatabasePassword, whatsSidDatabaseDialect));
 
-			WhatsSidService whatsSidService = new WhatsSidService(getEntityManager());
+				WhatsSidService whatsSidService = new WhatsSidService(getEntityManager());
 
-			if (Boolean.TRUE.equals(deleteAll)) {
-				deleteAllFingerprintings(whatsSidService);
+				if (Boolean.TRUE.equals(deleteAll)) {
+					deleteAllFingerprintings(whatsSidService);
+				}
 			}
-
 			if (directory != null) {
-				System.out.println(
-						"Create fingerprintings... (press q <return>, to abort after the current tune has been fingerprinted)");
+				System.out
+						.println("Create Recordings... (press q <return>, please wait for last recordings to finish)");
 
 				processDirectory(executor, directory);
 			}
@@ -155,7 +161,7 @@ public class FingerPrintingCreator {
 		} finally {
 			try {
 				executor.shutdown();
-				if (!executor.awaitTermination(FINGERPRINTING_AWAIT_TERMINATION, TimeUnit.DAYS)) {
+				if (!executor.awaitTermination(AWAIT_TERMINATION, TimeUnit.DAYS)) {
 					executor.shutdownNow();
 				}
 			} catch (InterruptedException e) {
@@ -200,7 +206,7 @@ public class FingerPrintingCreator {
 			if (file.isDirectory()) {
 				processDirectory(executor, file);
 			} else if (file.isFile()) {
-				if (!quit && TUNE_FILE_FILTER.accept(file)) {
+				if (!quit && (AUDIO_TUNE_FILE_FILTER.accept(file) || VIDEO_TUNE_FILE_FILTER.accept(file))) {
 					executor.execute(() -> {
 						try {
 							if (quit) {
@@ -226,37 +232,46 @@ public class FingerPrintingCreator {
 
 	private void processFile(ExecutorService executor, File file) throws IOException, SidTuneError {
 		try {
-			final WhatsSidService whatsSidService = new WhatsSidService(getEntityManager());
-
 			SidTune tune = SidTune.load(file);
 			String collectionName = PathUtils.getCollectionName(config.getSidplay2Section().getHvsc(), file);
 
-			WhatsSidDriver whatsSidDriver = new WhatsSidDriver();
-			whatsSidDriver
-					.setFingerprintInserter(new FingerPrinting(new IniFingerprintConfig(createIni), whatsSidService));
-			whatsSidDriver.setCollectionName(collectionName);
-			whatsSidDriver.setTune(tune);
-
 			Player player = new Player(config);
 			player.getC64().getVIC().setPalEmulation(PALEmulation.NONE);
-			player.setAudioDriver(whatsSidDriver);
 			player.setSidDatabase(new SidDatabase(config.getSidplay2Section().getHvsc()));
 
-			if (previousDirectory != null) {
-				copyRecordingsOfPreviousDirectory(player, whatsSidDriver, file, tune, collectionName);
+			AudioDriver audioDriver;
+			if (fingerprinting) {
+				final WhatsSidService whatsSidService = new WhatsSidService(getEntityManager());
+
+				WhatsSidDriver whatsSidDriver = new WhatsSidDriver();
+				whatsSidDriver.setFingerprintInserter(
+						new FingerPrinting(new IniFingerprintConfig(createIni), whatsSidService));
+				whatsSidDriver.setCollectionName(collectionName);
+				whatsSidDriver.setTune(tune);
+				audioDriver = whatsSidDriver;
+
+				if (previousDirectory != null) {
+					copyRecordingsOfPreviousDirectory(player, whatsSidDriver, file, tune, collectionName);
+				}
+			} else {
+				audioDriver = player.getConfig().getAudioSection().getAudio().newAudioDriver();
 			}
+			player.setAudioDriver(audioDriver);
+
 			player.setRecordingFilenameProvider(
 					theTune -> getRecordingFilename(file, theTune, theTune.getInfo().getCurrentSong()));
 			player.setTune(tune);
 			player.startC64();
 			player.stopC64(false);
 		} finally {
-			freeEntityManager();
+			if (fingerprinting) {
+				freeEntityManager();
+			}
 		}
 	}
 
-	private void copyRecordingsOfPreviousDirectory(Player player, WhatsSidDriver whatsSidDriver, File file,
-			SidTune tune, String collectionName) throws IOException, SidTuneError {
+	private void copyRecordingsOfPreviousDirectory(Player player, AudioDriver whatsSidDriver, File file, SidTune tune,
+			String collectionName) throws IOException, SidTuneError {
 		File previousFile = new File(previousDirectory, collectionName);
 		if (previousFile.exists()) {
 			SidTune previousTune = SidTune.load(previousFile);
@@ -265,12 +280,15 @@ public class FingerPrintingCreator {
 					&& player.getSidDatabaseInfo(db -> db.getTuneLength(tune), 0.) == previousSidDatabase
 							.getTuneLength(previousTune)) {
 				for (int songNo = 1; songNo <= tune.getInfo().getSongs(); songNo++) {
-					File wavFile = new File(getRecordingFilename(file, tune, songNo) + whatsSidDriver.getExtension());
-					File previousWavFile = new File(
+					File recordedFile = new File(
+							getRecordingFilename(file, tune, songNo) + whatsSidDriver.getExtension());
+					File previousRecordedFile = new File(
 							getRecordingFilename(previousFile, previousTune, songNo) + whatsSidDriver.getExtension());
-					if (!wavFile.exists() && previousWavFile.exists()) {
-						System.out.println(String.format("Tune is unchanged, copy %s to %s", previousWavFile, wavFile));
-						Files.copy(previousWavFile.toPath(), wavFile.toPath(), StandardCopyOption.REPLACE_EXISTING);
+					if (!recordedFile.exists() && previousRecordedFile.exists()) {
+						System.out.println(
+								String.format("Tune is unchanged, copy %s to %s", previousRecordedFile, recordedFile));
+						Files.copy(previousRecordedFile.toPath(), recordedFile.toPath(),
+								StandardCopyOption.REPLACE_EXISTING);
 					}
 				}
 			}
@@ -307,7 +325,7 @@ public class FingerPrintingCreator {
 	}
 
 	public static void main(String[] args) {
-		new FingerPrintingCreator().execute(args);
+		new RecordingTool().execute(args);
 	}
 
 }
